refactor(config): split runtime config seams
