+++ crates/app/src/config/runtime.rs
@@ -1,5 +1,5 @@
 use std::{
-    collections::{BTreeMap, BTreeSet},
+    collections::BTreeMap,
     fs,
     path::Path,
     path::PathBuf,
@@ -12,8 +12,31 @@ use serde::{Deserialize, Serialize};
 
 use crate::CliResult;
 use crate::mcp::McpConfig;
-use crate::secrets::DefaultSecretResolver;
-use loong_contracts::{SecretRef, SecretResolver};
+#[path = "runtime_provider_selector.rs"]
+mod provider_selector;
+#[path = "runtime_acp.rs"]
+mod runtime_acp;
+#[path = "runtime_provider_profiles.rs"]
+mod provider_profiles;
+
+pub use self::provider_selector::{
+    PROVIDER_SELECTOR_COMPACT_NOTE, PROVIDER_SELECTOR_HUMAN_SUMMARY, PROVIDER_SELECTOR_NOTE,
+    PROVIDER_SELECTOR_PLACEHOLDER, PROVIDER_SELECTOR_TARGET_SUMMARY, ProviderSelectorProfileRef,
+    ProviderSelectorResolution, accepted_provider_selectors, describe_provider_selector_target,
+    preferred_provider_selector, provider_selector_catalog,
+    provider_selector_recommendation_hint, resolve_provider_selector,
+};
+pub(crate) use self::runtime_acp::{normalize_dispatch_account_id, normalize_dispatch_channel_id};
+use self::provider_profiles::{
+    ActiveProviderSelectionBasis, ProviderSelectionNormalizationReport,
+    RawProviderSelectionIntent, normalize_provider_profile_id, parse_toml_config_components,
+    recover_active_provider_from_legacy_config,
+};
+pub use self::runtime_acp::{
+    AcpBackendProfilesConfig, AcpConfig, AcpConversationRoutingMode, AcpDispatchConfig,
+    AcpDispatchThreadRoutingMode, AcpxBackendConfig, AcpxMcpServerConfig, ControlPlaneConfig,
+    GatewayConfig,
+};
 
 use super::{
     OnebotChannelConfig, QqbotChannelConfig, WeixinChannelConfig,
@@ -31,7 +54,7 @@ use super::{
     feishu_integration::FeishuIntegrationConfig,
     memory::MemoryConfig,
     outbound_http::OutboundHttpConfig,
-    provider::{ProviderConfig, ProviderKind, ProviderProfileConfig},
+    provider::{ProviderConfig, ProviderProfileConfig},
     shared::{
         ConfigValidationIssue, ConfigValidationLocale, ConfigValidationSeverity,
         DEFAULT_CONFIG_FILE, default_loong_home as shared_default_loong_home, expand_path,
@@ -194,926 +217,6 @@ pub struct LoongConfig {
     pub control_plane: ControlPlaneConfig,
     #[serde(default)]
     pub acp: AcpConfig,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq)]
-pub struct GatewayConfig {
-    #[serde(default = "default_gateway_port")]
-    pub port: u16,
-}
-
-impl Default for GatewayConfig {
-    fn default() -> Self {
-        Self {
-            port: default_gateway_port(),
-        }
-    }
-}
-
-impl GatewayConfig {
-    fn is_default(config: &Self) -> bool {
-        *config == Self::default()
-    }
-}
-
-fn default_gateway_port() -> u16 {
-    26_306
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq, Default)]
-pub struct ControlPlaneConfig {
-    #[serde(default)]
-    pub allow_remote: bool,
-    #[serde(default)]
-    pub shared_token: Option<SecretRef>,
-}
-
-impl ControlPlaneConfig {
-    fn is_default(config: &Self) -> bool {
-        *config == Self::default()
-    }
-
-    pub fn resolved_shared_token(&self) -> CliResult<Option<String>> {
-        let Some(secret_ref) = self.shared_token.as_ref() else {
-            return Ok(None);
-        };
-        let resolver = DefaultSecretResolver::default();
-        let resolved_value = resolver
-            .resolve(secret_ref)
-            .map_err(|error| format!("resolve control-plane shared token failed: {error}"))?;
-        let shared_token = resolved_value.map(|secret| secret.into_inner());
-        Ok(shared_token)
-    }
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq)]
-pub struct AcpConfig {
-    #[serde(default)]
-    pub enabled: bool,
-    #[serde(default)]
-    pub backend: Option<String>,
-    #[serde(default)]
-    pub dispatch: AcpDispatchConfig,
-    #[serde(default)]
-    pub default_agent: Option<String>,
-    #[serde(default)]
-    pub allowed_agents: Vec<String>,
-    #[serde(default)]
-    pub max_concurrent_sessions: Option<usize>,
-    #[serde(default)]
-    pub session_idle_ttl_ms: Option<u64>,
-    #[serde(default)]
-    pub startup_timeout_ms: Option<u64>,
-    #[serde(default)]
-    pub turn_timeout_ms: Option<u64>,
-    #[serde(default)]
-    pub queue_owner_ttl_ms: Option<u64>,
-    #[serde(default)]
-    pub bindings_enabled: bool,
-    #[serde(default)]
-    pub emit_runtime_events: bool,
-    #[serde(default)]
-    pub allow_mcp_server_injection: bool,
-    #[serde(default)]
-    pub backends: AcpBackendProfilesConfig,
-}
-
-impl AcpConfig {
-    pub fn backend_id(&self) -> Option<String> {
-        self.backend
-            .as_deref()
-            .map(str::trim)
-            .filter(|value| !value.is_empty())
-            .map(|value| value.to_ascii_lowercase())
-    }
-
-    pub fn dispatch_enabled(&self) -> bool {
-        self.enabled && self.dispatch.enabled
-    }
-
-    pub fn max_concurrent_sessions(&self) -> usize {
-        self.max_concurrent_sessions
-            .filter(|value| *value > 0)
-            .unwrap_or_else(default_acp_max_concurrent_sessions)
-    }
-
-    pub fn resolved_default_agent(&self) -> CliResult<String> {
-        let raw = normalize_optional_string(self.default_agent.as_deref())
-            .unwrap_or_else(|| "codex".to_owned());
-        normalize_acp_agent_id(raw.as_str()).ok_or_else(|| {
-            format!("ACP default agent `{raw}` is invalid; use letters, numbers, `-`, or `_`")
-        })
-    }
-
-    pub fn allowed_agent_ids(&self) -> CliResult<Vec<String>> {
-        let default_agent = self.resolved_default_agent()?;
-        if self.allowed_agents.is_empty() {
-            return Ok(vec![default_agent]);
-        }
-
-        let mut seen = BTreeSet::new();
-        let mut agents = Vec::new();
-        for raw in &self.allowed_agents {
-            let trimmed = raw.trim();
-            let normalized = normalize_acp_agent_id(trimmed).ok_or_else(|| {
-                format!(
-                    "ACP allowed agent `{trimmed}` is invalid; use letters, numbers, `-`, or `_`"
-                )
-            })?;
-            if seen.insert(normalized.clone()) {
-                agents.push(normalized);
-            }
-        }
-
-        if !agents.iter().any(|agent| agent == &default_agent) {
-            return Err(format!(
-                "ACP default agent `{default_agent}` must be included in allowed_agents"
-            ));
-        }
-
-        Ok(agents)
-    }
-
-    pub fn resolve_allowed_agent(&self, raw: &str) -> CliResult<String> {
-        let normalized = normalize_acp_agent_id(raw).ok_or_else(|| {
-            format!("ACP agent `{raw}` is invalid; use letters, numbers, `-`, or `_`")
-        })?;
-        let allowed = self.allowed_agent_ids()?;
-        if allowed.iter().any(|agent| agent == &normalized) {
-            return Ok(normalized);
-        }
-        Err(format!(
-            "ACP agent `{normalized}` is not in the allowed ACP agents ({})",
-            allowed.join(", ")
-        ))
-    }
-
-    pub fn session_idle_ttl_ms(&self) -> u64 {
-        self.session_idle_ttl_ms
-            .filter(|value| *value > 0)
-            .unwrap_or_else(default_acp_session_idle_ttl_ms)
-    }
-
-    pub fn startup_timeout_ms(&self) -> u64 {
-        self.startup_timeout_ms
-            .filter(|value| *value > 0)
-            .unwrap_or_else(default_acp_startup_timeout_ms)
-    }
-
-    pub fn turn_timeout_ms(&self) -> u64 {
-        self.turn_timeout_ms
-            .filter(|value| *value > 0)
-            .unwrap_or_else(default_acp_turn_timeout_ms)
-    }
-
-    pub fn queue_owner_ttl_ms(&self) -> u64 {
-        self.queue_owner_ttl_ms
-            .filter(|value| *value > 0)
-            .unwrap_or_else(default_acp_queue_owner_ttl_ms)
-    }
-
-    pub fn acpx_profile(&self) -> Option<&AcpxBackendConfig> {
-        self.backends.acpx.as_ref()
-    }
-}
-
-impl Default for AcpConfig {
-    fn default() -> Self {
-        Self {
-            enabled: false,
-            backend: None,
-            dispatch: AcpDispatchConfig::default(),
-            default_agent: None,
-            allowed_agents: Vec::new(),
-            max_concurrent_sessions: Some(default_acp_max_concurrent_sessions()),
-            session_idle_ttl_ms: Some(default_acp_session_idle_ttl_ms()),
-            startup_timeout_ms: Some(default_acp_startup_timeout_ms()),
-            turn_timeout_ms: Some(default_acp_turn_timeout_ms()),
-            queue_owner_ttl_ms: Some(default_acp_queue_owner_ttl_ms()),
-            bindings_enabled: false,
-            emit_runtime_events: false,
-            allow_mcp_server_injection: false,
-            backends: AcpBackendProfilesConfig::default(),
-        }
-    }
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq)]
-pub struct AcpDispatchConfig {
-    #[serde(default = "default_true")]
-    pub enabled: bool,
-    #[serde(default)]
-    pub conversation_routing: AcpConversationRoutingMode,
-    #[serde(default)]
-    pub allowed_channels: Vec<String>,
-    #[serde(default)]
-    pub allowed_account_ids: Vec<String>,
-    #[serde(default)]
-    pub bootstrap_mcp_servers: Vec<String>,
-    #[serde(default)]
-    pub working_directory: Option<String>,
-    #[serde(default)]
-    pub thread_routing: AcpDispatchThreadRoutingMode,
-}
-
-impl Default for AcpDispatchConfig {
-    fn default() -> Self {
-        Self {
-            enabled: default_true(),
-            conversation_routing: AcpConversationRoutingMode::default(),
-            allowed_channels: Vec::new(),
-            allowed_account_ids: Vec::new(),
-            bootstrap_mcp_servers: Vec::new(),
-            working_directory: None,
-            thread_routing: AcpDispatchThreadRoutingMode::default(),
-        }
-    }
-}
-
-impl AcpDispatchConfig {
-    pub fn allowed_channel_ids(&self) -> CliResult<Vec<String>> {
-        let mut seen = BTreeSet::new();
-        let mut channels = Vec::new();
-        for raw in &self.allowed_channels {
-            let trimmed = raw.trim();
-            let normalized = normalize_dispatch_channel_id(trimmed).ok_or_else(|| {
-                format!(
-                    "ACP dispatch allowed channel `{trimmed}` is invalid; use letters, numbers, `-`, or `_`"
-                )
-            })?;
-            if seen.insert(normalized.clone()) {
-                channels.push(normalized);
-            }
-        }
-        Ok(channels)
-    }
-
-    pub fn allows_channel_id(&self, channel_id: Option<&str>) -> CliResult<bool> {
-        let allowed = self.allowed_channel_ids()?;
-        if allowed.is_empty() {
-            return Ok(true);
-        }
-        let Some(channel_id) = channel_id.and_then(normalize_dispatch_channel_id) else {
-            return Ok(false);
-        };
-        Ok(allowed.iter().any(|channel| channel == &channel_id))
-    }
-
-    pub fn allowed_account_ids(&self) -> CliResult<Vec<String>> {
-        let mut seen = BTreeSet::new();
-        let mut accounts = Vec::new();
-        for raw in &self.allowed_account_ids {
-            let trimmed = raw.trim();
-            let normalized = normalize_dispatch_account_id(trimmed).ok_or_else(|| {
-                format!(
-                    "ACP dispatch allowed account `{trimmed}` is invalid; use a configured account identity or label"
-                )
-            })?;
-            if seen.insert(normalized.clone()) {
-                accounts.push(normalized);
-            }
-        }
-        Ok(accounts)
-    }
-
-    pub fn bootstrap_mcp_server_names(&self) -> CliResult<Vec<String>> {
-        self.bootstrap_mcp_server_names_with_additions(&[])
-    }
-
-    pub fn resolved_working_directory(&self) -> Option<PathBuf> {
-        self.working_directory
-            .as_deref()
-            .map(str::trim)
-            .filter(|value| !value.is_empty())
-            .map(Path::new)
-            .map(Path::to_path_buf)
-    }
-
-    pub fn bootstrap_mcp_server_names_with_additions(
-        &self,
-        additional: &[String],
-    ) -> CliResult<Vec<String>> {
-        let mut seen = BTreeSet::new();
-        let mut names = Vec::new();
-        for raw in self
-            .bootstrap_mcp_servers
-            .iter()
-            .map(String::as_str)
-            .chain(additional.iter().map(String::as_str))
-        {
-            let Some(normalized) =
-                normalize_optional_string(Some(raw)).map(|value| value.to_ascii_lowercase())
-            else {
-                return Err(
-                    "ACP dispatch bootstrap MCP server names must not contain empty entries"
-                        .to_owned(),
-                );
-            };
-            if seen.insert(normalized.clone()) {
-                names.push(normalized);
-            }
-        }
-        Ok(names)
-    }
-
-    pub fn allows_account_id(&self, account_id: Option<&str>) -> CliResult<bool> {
-        let allowed = self.allowed_account_ids()?;
-        if allowed.is_empty() {
-            return Ok(true);
-        }
-        let Some(account_id) = account_id.and_then(normalize_dispatch_account_id) else {
-            return Ok(false);
-        };
-        Ok(allowed.iter().any(|candidate| candidate == &account_id))
-    }
-
-    pub fn allows_thread_id(&self, thread_id: Option<&str>) -> bool {
-        self.thread_routing.allows_thread_id(thread_id)
-    }
-}
-
-#[derive(Debug, Clone, Copy, Serialize, Deserialize, PartialEq, Eq, Default)]
-#[serde(rename_all = "snake_case")]
-pub enum AcpConversationRoutingMode {
-    #[default]
-    AgentPrefixedOnly,
-    All,
-}
-
-impl AcpConversationRoutingMode {
-    pub const fn as_str(self) -> &'static str {
-        match self {
-            Self::All => "all",
-            Self::AgentPrefixedOnly => "agent_prefixed_only",
-        }
-    }
-}
-
-#[derive(Debug, Clone, Copy, Serialize, Deserialize, PartialEq, Eq, Default)]
-#[serde(rename_all = "snake_case")]
-pub enum AcpDispatchThreadRoutingMode {
-    #[default]
-    All,
-    ThreadOnly,
-    RootOnly,
-}
-
-impl AcpDispatchThreadRoutingMode {
-    pub const fn as_str(self) -> &'static str {
-        match self {
-            Self::All => "all",
-            Self::ThreadOnly => "thread_only",
-            Self::RootOnly => "root_only",
-        }
-    }
-
-    pub fn allows_thread_id(self, thread_id: Option<&str>) -> bool {
-        let has_thread = thread_id
-            .map(str::trim)
-            .filter(|value| !value.is_empty())
-            .is_some();
-        match self {
-            Self::All => true,
-            Self::ThreadOnly => has_thread,
-            Self::RootOnly => !has_thread,
-        }
-    }
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize, Default, PartialEq)]
-pub struct AcpBackendProfilesConfig {
-    #[serde(default)]
-    pub acpx: Option<AcpxBackendConfig>,
-}
-
-#[derive(Debug, Clone, Default, Serialize, Deserialize, PartialEq)]
-pub struct AcpxBackendConfig {
-    #[serde(default)]
-    pub command: Option<String>,
-    #[serde(default)]
-    pub expected_version: Option<String>,
-    #[serde(default)]
-    pub cwd: Option<String>,
-    #[serde(default)]
-    pub permission_mode: Option<String>,
-    #[serde(default)]
-    pub non_interactive_permissions: Option<String>,
-    #[serde(default)]
-    pub strict_windows_cmd_wrapper: Option<bool>,
-    #[serde(default)]
-    pub timeout_seconds: Option<f64>,
-    #[serde(default)]
-    pub queue_owner_ttl_seconds: Option<f64>,
-    #[serde(default)]
-    pub mcp_servers: BTreeMap<String, AcpxMcpServerConfig>,
-}
-
-impl AcpxBackendConfig {
-    pub fn command(&self) -> Option<String> {
-        normalize_optional_string(self.command.as_deref())
-    }
-
-    pub fn expected_version(&self) -> Option<String> {
-        normalize_optional_string(self.expected_version.as_deref())
-    }
-
-    pub fn cwd(&self) -> Option<String> {
-        normalize_optional_string(self.cwd.as_deref())
-    }
-
-    pub fn permission_mode(&self) -> Option<String> {
-        normalize_optional_string(self.permission_mode.as_deref())
-    }
-
-    pub fn non_interactive_permissions(&self) -> Option<String> {
-        normalize_optional_string(self.non_interactive_permissions.as_deref())
-    }
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq)]
-pub struct AcpxMcpServerConfig {
-    pub command: String,
-    #[serde(default)]
-    pub args: Vec<String>,
-    #[serde(default)]
-    pub env: BTreeMap<String, String>,
-}
-
-const fn default_acp_max_concurrent_sessions() -> usize {
-    8
-}
-
-const fn default_true() -> bool {
-    true
-}
-
-const fn default_acp_session_idle_ttl_ms() -> u64 {
-    900_000
-}
-
-const fn default_acp_startup_timeout_ms() -> u64 {
-    15_000
-}
-
-const fn default_acp_turn_timeout_ms() -> u64 {
-    120_000
-}
-
-const fn default_acp_queue_owner_ttl_ms() -> u64 {
-    30_000
-}
-
-fn normalize_optional_string(raw: Option<&str>) -> Option<String> {
-    raw.map(str::trim)
-        .filter(|value| !value.is_empty())
-        .map(|value| value.to_owned())
-}
-
-fn normalize_provider_profile_id(raw: &str) -> Option<String> {
-    normalize_dispatch_channel_id(raw)
-}
-
-#[derive(Debug, Clone, Default)]
-struct RawProviderSelectionIntent {
-    legacy_provider_explicit: bool,
-    active_provider_explicit: bool,
-    raw_active_provider: Option<String>,
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-enum ActiveProviderSelectionBasis {
-    ExplicitActiveProvider,
-    ExplicitLegacyProvider,
-    FirstSavedProfile,
-    LegacyOnly,
-}
-
-impl ActiveProviderSelectionBasis {
-    const fn diagnostic_summary(self) -> &'static str {
-        match self {
-            Self::ExplicitActiveProvider => "the explicit active_provider value",
-            Self::ExplicitLegacyProvider => "the explicit legacy [provider] table",
-            Self::FirstSavedProfile => "the first saved provider profile in sorted order",
-            Self::LegacyOnly => "the legacy [provider] table",
-        }
-    }
-}
-
-#[derive(Debug, Clone, Default)]
-struct ProviderSelectionNormalizationReport {
-    legacy_provider_explicit: bool,
-    active_provider_explicit: bool,
-    requested_active_provider: Option<String>,
-    selected_active_provider: Option<String>,
-    configured_profile_ids: Vec<String>,
-    selection_basis: Option<ActiveProviderSelectionBasis>,
-    warn_implicit_active_provider: bool,
-    warn_unknown_active_provider: bool,
-    recovered_legacy_profile_id: Option<String>,
-    legacy_profile_inserted: bool,
-    legacy_provider_validation_issues: Vec<ConfigValidationIssue>,
-}
-
-impl ProviderSelectionNormalizationReport {
-    fn validation_issues(&self) -> Vec<ConfigValidationIssue> {
-        let mut issues = self.legacy_provider_validation_issues.clone();
-        let configured_profile_ids = self.configured_profile_ids.join(", ");
-        let selected_profile_id = self
-            .selected_active_provider
-            .as_deref()
-            .unwrap_or("unknown")
-            .to_owned();
-        let selection_basis = self
-            .selection_basis
-            .map(ActiveProviderSelectionBasis::diagnostic_summary)
-            .unwrap_or("provider profile normalization")
-            .to_owned();
-
-        if self.warn_implicit_active_provider {
-            let mut extra_message_variables = BTreeMap::new();
-            extra_message_variables.insert("selected_profile_id".to_owned(), selected_profile_id);
-            extra_message_variables.insert("selection_basis".to_owned(), selection_basis.clone());
-            extra_message_variables.insert(
-                "configured_profile_ids".to_owned(),
-                configured_profile_ids.clone(),
-            );
-            issues.push(ConfigValidationIssue {
-                severity: ConfigValidationSeverity::Warn,
-                code: super::shared::ConfigValidationCode::ImplicitActiveProvider,
-                field_path: "active_provider".to_owned(),
-                inline_field_path: "providers".to_owned(),
-                example_env_name: String::new(),
-                suggested_env_name: self.selected_active_provider.clone(),
-                extra_message_variables,
-            });
-        }
-
-        if self.warn_unknown_active_provider {
-            let mut extra_message_variables = BTreeMap::new();
-            extra_message_variables.insert(
-                "requested_profile_id".to_owned(),
-                self.requested_active_provider
-                    .as_deref()
-                    .map(str::trim)
-                    .filter(|value| !value.is_empty())
-                    .unwrap_or("(blank)")
-                    .to_owned(),
-            );
-            extra_message_variables.insert(
-                "selected_profile_id".to_owned(),
-                self.selected_active_provider
-                    .clone()
-                    .unwrap_or_else(|| "unknown".to_owned()),
-            );
-            extra_message_variables.insert("selection_basis".to_owned(), selection_basis);
-            extra_message_variables
-                .insert("configured_profile_ids".to_owned(), configured_profile_ids);
-            issues.push(ConfigValidationIssue {
-                severity: ConfigValidationSeverity::Warn,
-                code: super::shared::ConfigValidationCode::UnknownActiveProvider,
-                field_path: "active_provider".to_owned(),
-                inline_field_path: "providers".to_owned(),
-                example_env_name: String::new(),
-                suggested_env_name: self.selected_active_provider.clone(),
-                extra_message_variables,
-            });
-        }
-
-        issues
-    }
-}
-
-pub const PROVIDER_SELECTOR_PLACEHOLDER: &str = "<profile|model|kind>";
-pub const PROVIDER_SELECTOR_HUMAN_SUMMARY: &str =
-    "profile id, unique model name or suffix, or provider kind";
-pub const PROVIDER_SELECTOR_TARGET_SUMMARY: &str =
-    "target profile id, unique model name or suffix, or provider kind";
-pub const PROVIDER_SELECTOR_NOTE: &str =
-    "you can also enter a unique model name, model suffix, or provider kind";
-pub const PROVIDER_SELECTOR_COMPACT_NOTE: &str = "type a model, suffix, or provider kind";
-
-fn normalize_provider_selector_token(raw: &str) -> Option<String> {
-    let trimmed = raw.trim();
-    (!trimmed.is_empty()).then(|| trimmed.to_ascii_lowercase())
-}
-
-fn provider_model_suffix(raw: &str) -> Option<String> {
-    let normalized = normalize_provider_selector_token(raw)?;
-    normalized
-        .rsplit('/')
-        .next()
-        .map(str::trim)
-        .filter(|value| !value.is_empty())
-        .map(str::to_owned)
-}
-
-fn push_unique_selector(selectors: &mut Vec<String>, candidate: &str) {
-    if candidate.trim().is_empty() {
-        return;
-    }
-    if selectors
-        .iter()
-        .any(|existing| existing.eq_ignore_ascii_case(candidate))
-    {
-        return;
-    }
-    selectors.push(candidate.to_owned());
-}
-
-#[derive(Debug, Clone, Copy)]
-pub struct ProviderSelectorProfileRef<'a> {
-    pub profile_id: &'a str,
-    pub kind: ProviderKind,
-    pub model: &'a str,
-    pub default_for_kind: bool,
-}
-
-impl<'a> ProviderSelectorProfileRef<'a> {
-    pub const fn new(
-        profile_id: &'a str,
-        kind: ProviderKind,
-        model: &'a str,
-        default_for_kind: bool,
-    ) -> Self {
-        Self {
-            profile_id,
-            kind,
-            model,
-            default_for_kind,
-        }
-    }
-}
-
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub enum ProviderSelectorResolution {
-    Match(String),
-    Ambiguous(Vec<String>),
-    NoMatch,
-}
-
-pub fn accepted_provider_selectors<'a, I>(profiles: I, target_profile_id: &str) -> Vec<String>
-where
-    I: IntoIterator<Item = ProviderSelectorProfileRef<'a>>,
-{
-    ProviderSelectorIndex::new(profiles).accepted_selectors(target_profile_id)
-}
-
-pub fn provider_selector_catalog<'a, I>(profiles: I) -> Vec<String>
-where
-    I: IntoIterator<Item = ProviderSelectorProfileRef<'a>>,
-{
-    ProviderSelectorIndex::new(profiles).selector_catalog()
-}
-
-pub fn preferred_provider_selector<'a, I>(profiles: I, target_profile_id: &str) -> Option<String>
-where
-    I: IntoIterator<Item = ProviderSelectorProfileRef<'a>>,
-{
-    ProviderSelectorIndex::new(profiles).preferred_selector(target_profile_id)
-}
-
-pub fn describe_provider_selector_target<'a, I>(
-    profiles: I,
-    target_profile_id: &str,
-) -> Option<String>
-where
-    I: IntoIterator<Item = ProviderSelectorProfileRef<'a>>,
-{
-    ProviderSelectorIndex::new(profiles).describe_profile(target_profile_id)
-}
-
-pub fn provider_selector_recommendation_hint<'a, I, J, S>(
-    profiles: I,
-    target_profile_ids: J,
-) -> Option<String>
-where
-    I: IntoIterator<Item = ProviderSelectorProfileRef<'a>>,
-    J: IntoIterator<Item = S>,
-    S: AsRef<str>,
-{
-    ProviderSelectorIndex::new(profiles).recommendation_hint(target_profile_ids)
-}
-
-pub fn resolve_provider_selector<'a, I>(profiles: I, selector: &str) -> ProviderSelectorResolution
-where
-    I: IntoIterator<Item = ProviderSelectorProfileRef<'a>>,
-{
-    ProviderSelectorIndex::new(profiles).resolve(selector)
-}
-
-struct ProviderSelectorIndex<'a> {
-    profiles: Vec<ProviderSelectorProfileRef<'a>>,
-}
-
-impl<'a> ProviderSelectorIndex<'a> {
-    fn new<I>(profiles: I) -> Self
-    where
-        I: IntoIterator<Item = ProviderSelectorProfileRef<'a>>,
-    {
-        Self {
-            profiles: profiles.into_iter().collect(),
-        }
-    }
-
-    fn accepted_selectors(&self, target_profile_id: &str) -> Vec<String> {
-        let Some(profile) = self.find_profile(target_profile_id) else {
-            return Vec::new();
-        };
-
-        let mut selectors = Vec::new();
-        push_unique_selector(&mut selectors, profile.profile_id);
-
-        if let Some(model) = normalize_provider_selector_token(profile.model)
-            && self.model_matches(model.as_str()).len() == 1
-        {
-            push_unique_selector(&mut selectors, model.as_str());
-        }
-
-        if let Some(suffix) = provider_model_suffix(profile.model)
-            && self.model_suffix_matches(suffix.as_str()).len() == 1
-        {
-            push_unique_selector(&mut selectors, suffix.as_str());
-        }
-
-        if self.kind_resolves_to_profile_id(profile.kind, profile.profile_id) {
-            push_unique_selector(&mut selectors, profile.kind.as_str());
-        }
-
-        selectors
-    }
-
-    fn resolve(&self, selector: &str) -> ProviderSelectorResolution {
-        let Some(normalized) = normalize_provider_selector_token(selector) else {
-            return ProviderSelectorResolution::NoMatch;
-        };
-
-        if let Some(profile) = self
-            .profiles
-            .iter()
-            .find(|profile| profile.profile_id.eq_ignore_ascii_case(normalized.as_str()))
-        {
-            return ProviderSelectorResolution::Match(profile.profile_id.to_owned());
-        }
-
-        let model_matches = self.model_matches(normalized.as_str());
-        match model_matches.as_slice() {
-            [profile_id] => return ProviderSelectorResolution::Match(profile_id.clone()),
-            matches if matches.len() > 1 => {
-                return ProviderSelectorResolution::Ambiguous(matches.to_vec());
-            }
-            _ => {}
-        }
-
-        let suffix_matches = self.model_suffix_matches(normalized.as_str());
-        match suffix_matches.as_slice() {
-            [profile_id] => return ProviderSelectorResolution::Match(profile_id.clone()),
-            matches if matches.len() > 1 => {
-                return ProviderSelectorResolution::Ambiguous(matches.to_vec());
-            }
-            _ => {}
-        }
-
-        let Some(kind) = ProviderKind::parse(normalized.as_str()) else {
-            return ProviderSelectorResolution::NoMatch;
-        };
-        self.resolve_kind(kind)
-    }
-
-    fn selector_catalog(&self) -> Vec<String> {
-        let mut selectors = Vec::new();
-        for profile in &self.profiles {
-            for selector in self.accepted_selectors(profile.profile_id) {
-                push_unique_selector(&mut selectors, selector.as_str());
-            }
-        }
-        selectors
-    }
-
-    fn preferred_selector(&self, target_profile_id: &str) -> Option<String> {
-        let profile = self.find_profile(target_profile_id)?;
-        let selectors = self.accepted_selectors(target_profile_id);
-        if selectors.is_empty() {
-            return None;
-        }
-
-        let profile_id = normalize_provider_selector_token(profile.profile_id);
-        let model = normalize_provider_selector_token(profile.model);
-        let suffix = provider_model_suffix(profile.model);
-        let kind = normalize_provider_selector_token(profile.kind.as_str());
-        let profile_id_len = profile_id.as_ref().map_or(usize::MAX, String::len);
-
-        let preferred_candidates = [
-            kind.as_deref(),
-            suffix.as_deref(),
-            model
-                .as_deref()
-                .filter(|model| model.len() <= profile_id_len),
-            profile_id.as_deref(),
-            model.as_deref(),
-        ];
-
-        for candidate in preferred_candidates.into_iter().flatten() {
-            if let Some(selector) = selectors
-                .iter()
-                .find(|existing| existing.eq_ignore_ascii_case(candidate))
-            {
-                return Some(selector.clone());
-            }
-        }
-
-        selectors.into_iter().next()
-    }
-
-    fn describe_profile(&self, target_profile_id: &str) -> Option<String> {
-        let profile = self.find_profile(target_profile_id)?;
-        let selectors = self.accepted_selectors(target_profile_id);
-        let mut description = format!("{} [model={}", profile.profile_id, profile.model);
-        if !selectors.is_empty() {
-            description.push_str("; selectors=");
-            description.push_str(selectors.join(", ").as_str());
-        }
-        description.push(']');
-        Some(description)
-    }
-
-    fn recommendation_hint<J, S>(&self, target_profile_ids: J) -> Option<String>
-    where
-        J: IntoIterator<Item = S>,
-        S: AsRef<str>,
-    {
-        let mut selectors = Vec::new();
-        for profile_id in target_profile_ids {
-            let Some(selector) = self.preferred_selector(profile_id.as_ref()) else {
-                continue;
-            };
-            push_unique_selector(&mut selectors, selector.as_str());
-            if selectors.len() >= 3 {
-                break;
-            }
-        }
-        (!selectors.is_empty()).then(|| format!("try one of: {}", selectors.join(", ")))
-    }
-
-    fn find_profile(&self, profile_id: &str) -> Option<&ProviderSelectorProfileRef<'a>> {
-        self.profiles
-            .iter()
-            .find(|profile| profile.profile_id == profile_id)
-    }
-
-    fn model_matches(&self, selector: &str) -> Vec<String> {
-        self.profiles
-            .iter()
-            .filter(|profile| {
-                normalize_provider_selector_token(profile.model).as_deref() == Some(selector)
-            })
-            .map(|profile| profile.profile_id.to_owned())
-            .collect()
-    }
-
-    fn model_suffix_matches(&self, selector: &str) -> Vec<String> {
-        self.profiles
-            .iter()
-            .filter(|profile| provider_model_suffix(profile.model).as_deref() == Some(selector))
-            .map(|profile| profile.profile_id.to_owned())
-            .collect()
-    }
-
-    fn resolve_kind(&self, kind: ProviderKind) -> ProviderSelectorResolution {
-        let matches = self
-            .profiles
-            .iter()
-            .filter(|profile| profile.kind == kind)
-            .collect::<Vec<_>>();
-        let Some(first) = matches.first().copied() else {
-            return ProviderSelectorResolution::NoMatch;
-        };
-        if matches.len() == 1 {
-            return ProviderSelectorResolution::Match(first.profile_id.to_owned());
-        }
-
-        let default_matches = matches
-            .iter()
-            .copied()
-            .filter(|profile| profile.default_for_kind)
-            .collect::<Vec<_>>();
-        if let [default_match] = default_matches.as_slice() {
-            return ProviderSelectorResolution::Match(default_match.profile_id.to_owned());
-        }
-
-        ProviderSelectorResolution::Ambiguous(
-            matches
-                .into_iter()
-                .map(|profile| profile.profile_id.to_owned())
-                .collect(),
-        )
-    }
-
-    fn kind_resolves_to_profile_id(&self, kind: ProviderKind, profile_id: &str) -> bool {
-        matches!(
-            self.resolve_kind(kind),
-            ProviderSelectorResolution::Match(resolved) if resolved == profile_id
-        )
-    }
 }
 
 fn canonicalize_provider_profile_for_encoding(profile: &mut ProviderProfileConfig) {
@@ -1511,72 +614,6 @@ fn canonicalize_env_string_reference(value: &mut Option<String>, env_name: &mut 
             *value = None;
         }
         Some(_) => {}
-    }
-}
-
-fn normalize_acp_agent_id(raw: &str) -> Option<String> {
-    let normalized = raw.trim().to_ascii_lowercase();
-    let mut chars = normalized.chars();
-    let first = chars.next()?;
-    if !first.is_ascii_alphanumeric() {
-        return None;
-    }
-    if !chars.all(|ch| ch.is_ascii_alphanumeric() || ch == '-' || ch == '_') {
-        return None;
-    }
-    Some(normalized)
-}
-
-pub(crate) fn normalize_dispatch_channel_id(raw: &str) -> Option<String> {
-    let normalized = raw.trim().to_ascii_lowercase();
-    let mut chars = normalized.chars();
-    let first = chars.next()?;
-    if !first.is_ascii_alphanumeric() {
-        return None;
-    }
-    if !chars.all(|ch| ch.is_ascii_alphanumeric() || ch == '-' || ch == '_') {
-        return None;
-    }
-    Some(normalized)
-}
-
-pub(crate) fn normalize_dispatch_account_id(raw: &str) -> Option<String> {
-    let trimmed = raw.trim();
-    if trimmed.is_empty() {
-        return None;
-    }
-
-    let mut normalized = String::with_capacity(trimmed.len());
-    let mut last_was_separator = false;
-    let mut saw_alphanumeric = false;
-    for value in trimmed.chars() {
-        if value.is_ascii_alphanumeric() {
-            normalized.push(value.to_ascii_lowercase());
-            last_was_separator = false;
-            saw_alphanumeric = true;
-            continue;
-        }
-        if matches!(value, '_' | '-') {
-            if !normalized.is_empty() && !last_was_separator {
-                normalized.push(value);
-                last_was_separator = true;
-            }
-            continue;
-        }
-        if !normalized.is_empty() && !last_was_separator {
-            normalized.push('-');
-            last_was_separator = true;
-        }
-    }
-
-    while matches!(normalized.chars().last(), Some('-' | '_')) {
-        normalized.pop();
-    }
-
-    if !saw_alphanumeric || normalized.is_empty() {
-        None
-    } else {
-        Some(normalized)
     }
 }
 
@@ -1981,111 +1018,6 @@ impl LoongConfig {
         canonicalize_channel_configs_for_encoding(&mut cloned);
         cloned
     }
-}
-
-fn normalized_inferred_profile_id(provider: &ProviderConfig) -> String {
-    normalize_provider_profile_id(provider.inferred_profile_id().as_str())
-        .unwrap_or_else(|| provider.inferred_profile_id())
-}
-
-fn matching_legacy_provider_profile_id(
-    providers: &BTreeMap<String, ProviderProfileConfig>,
-    legacy_provider: &ProviderConfig,
-) -> Option<String> {
-    let inferred_profile_id = normalized_inferred_profile_id(legacy_provider);
-    if providers
-        .get(&inferred_profile_id)
-        .is_some_and(|profile| profile.provider == *legacy_provider)
-    {
-        return Some(inferred_profile_id);
-    }
-
-    let exact_matches = providers
-        .iter()
-        .filter(|(_profile_id, profile)| profile.provider == *legacy_provider)
-        .map(|(profile_id, _profile)| profile_id.clone())
-        .collect::<Vec<_>>();
-    if exact_matches.len() == 1 {
-        return exact_matches.into_iter().next();
-    }
-    exact_matches.into_iter().next()
-}
-
-fn next_available_provider_profile_id(
-    providers: &BTreeMap<String, ProviderProfileConfig>,
-    base_profile_id: &str,
-) -> String {
-    if !providers.contains_key(base_profile_id) {
-        return base_profile_id.to_owned();
-    }
-    let max_suffix = providers.len().saturating_add(2);
-    for suffix in 2..=max_suffix {
-        let candidate = format!("{base_profile_id}-{suffix}");
-        if !providers.contains_key(&candidate) {
-            return candidate;
-        }
-    }
-    format!("{base_profile_id}-{max_suffix}")
-}
-
-fn recover_active_provider_from_legacy_config(
-    legacy_provider: &ProviderConfig,
-    providers: &mut BTreeMap<String, ProviderProfileConfig>,
-) -> (String, bool) {
-    if let Some(profile_id) = matching_legacy_provider_profile_id(providers, legacy_provider) {
-        return (profile_id, false);
-    }
-
-    let profile_id = next_available_provider_profile_id(
-        providers,
-        normalized_inferred_profile_id(legacy_provider).as_str(),
-    );
-    let mut recovered_profile = ProviderProfileConfig::from_provider(legacy_provider.clone());
-    recovered_profile.default_for_kind = !providers
-        .values()
-        .any(|profile| profile.provider.kind == recovered_profile.provider.kind);
-    providers.insert(profile_id.clone(), recovered_profile);
-    (profile_id, true)
-}
-
-#[cfg(feature = "config-toml")]
-fn inspect_raw_provider_selection_intent(raw: &str) -> CliResult<RawProviderSelectionIntent> {
-    let value = toml::from_str::<toml::Value>(raw)
-        .map_err(|error| format!("failed to parse TOML config: {error}"))?;
-    let table = value.as_table();
-    Ok(RawProviderSelectionIntent {
-        legacy_provider_explicit: table.is_some_and(|root| root.contains_key("provider")),
-        active_provider_explicit: table.is_some_and(|root| root.contains_key("active_provider")),
-        raw_active_provider: table
-            .and_then(|root| root.get("active_provider"))
-            .and_then(toml::Value::as_str)
-            .map(str::to_owned),
-    })
-}
-
-#[cfg(feature = "config-toml")]
-fn parse_toml_config_components(
-    raw: &str,
-) -> CliResult<(LoongConfig, ProviderSelectionNormalizationReport)> {
-    let mut config = toml::from_str::<LoongConfig>(raw)
-        .map_err(|error| format!("failed to parse TOML config: {error}"))?;
-    let selection_intent = inspect_raw_provider_selection_intent(raw)?;
-    let had_saved_provider_profiles = !config.providers.is_empty();
-    let legacy_provider_before_normalization = config.provider.clone();
-    let mut selection_report =
-        config.normalize_provider_profiles_with_intent(Some(&selection_intent));
-    if selection_intent.legacy_provider_explicit && had_saved_provider_profiles {
-        selection_report.legacy_provider_validation_issues =
-            legacy_provider_before_normalization.validate();
-    }
-    Ok((config, selection_report))
-}
-
-#[cfg(not(feature = "config-toml"))]
-fn parse_toml_config_components(
-    _raw: &str,
-) -> CliResult<(LoongConfig, ProviderSelectionNormalizationReport)> {
-    Err("config-toml feature is disabled for this build".to_owned())
 }
 
 pub fn load(path: Option<&str>) -> CliResult<(PathBuf, LoongConfig)> {
